# local-test/reference/notification/api.yaml
openapi: 3.0.0
info:
    title: Notification Management API
    version: 1.0.0
    description: API for managing user notifications

paths:
    /notifications:
        get:
            summary: Get notifications
            operationId: GetNotifications
            parameters:
                - name: limit
                  in: query
                  schema:
                      type: integer
                      format: int32
                  required: true
                - name: offset
                  in: query
                  schema:
                      type: integer
                      format: int32
                  required: true
            responses:
                "200":
                    description: Notifications retrieved successfully
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Notifications"
                "400":
                    description: Bad request
                "500":
                    description: Internal server error

    /notifications/unread:
        get:
            summary: Get unread notifications
            operationId: GetUnreadNotifications
            parameters:
                - name: limit
                  in: query
                  schema:
                      type: integer
                      format: int32
                  required: true
                - name: offset
                  in: query
                  schema:
                      type: integer
                      format: int32
                  required: true
            responses:
                "200":
                    description: Unread notifications retrieved successfully
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Notifications"
                "400":
                    description: Bad request
                "500":
                    description: Internal server error

    /notifications/unread/count:
        get:
            summary: Get unread notifications count
            operationId: GetUnreadNotificationsCount
            responses:
                "200":
                    description: Unread notifications count retrieved successfully
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Count"
                "400":
                    description: Bad request
                "500":
                    description: Internal server error

    /notifications/{notification_id}/read:
        patch:
            summary: Mark a notification as read
            operationId: MarkNotificationAsRead
            parameters:
                - name: notification_id
                  in: path
                  required: true
                  schema:
                      type: integer
                      format: int64
            responses:
                "200":
                    description: Notification marked as read successfully
                "400":
                    description: Bad request
                "500":
                    description: Internal server error
    /notifications/read/all:
        patch:
            summary: Mark all notifications as read
            operationId: MarkAllNotificationsAsRead
            responses:
                "200":
                    description: All notifications marked as read successfully
                "400":
                    description: Bad request
                "500":
                    description: Internal server error

components:
    schemas:
        Count:
            type: object
            required:
                - count
            properties:
                count:
                    type: integer
                    format: int64
        UserInfo:
            type: object
            properties:
                user_id:
                    type: string
                user_name:
                    type: string
                bio:
                    type: string
                profile_image_url:
                    type: string
                    format: uri
                is_private:
                    type: boolean
                is_admin:
                    type: boolean
                is_following:
                    type: boolean
                is_followed:
                    type: boolean
                is_pending:
                    type: boolean
            required:
                - user_id
                - user_name
                - bio
                - profile_image_url
                - is_private
                - is_admin
                - is_following
                - is_followed
                - is_pending
        Notifications:
            type: array
            items:
                $ref: "#/components/schemas/Notification"
        Notification:
            type: object
            required:
                - id
                - type
                - is_read
                - created_at
            properties:
                id:
                    type: integer
                    format: int64
                sender_info:
                    $ref: "#/components/schemas/UserInfo"
                    description: The information of the sender
                type:
                    type: string
                content:
                    type: string
                related_tweet:
                    $ref: "#/components/schemas/TweetInfo"
                    nullable: true
                is_read:
                    type: boolean
                created_at:
                    type: string
                    format: date-time
        TweetInfo:
            type: object
            properties:
                tweet_id:
                    type: integer
                    format: int64
                user_info:
                    $ref: "#/components/schemas/UserInfoWithoutBio"
                content:
                    type: string
                    nullable: true
                code:
                    type: string
                    nullable: true
                media:
                    $ref: "#/components/schemas/Media"
                    nullable: true
                likes_count:
                    type: integer
                    format: int32
                retweets_count:
                    type: integer
                    format: int32
                replies_count:
                    type: integer
                    format: int32
                is_quote:
                    type: boolean
                is_reply:
                    type: boolean
                is_pinned:
                    type: boolean
                has_liked:
                    type: boolean
                has_retweeted:
                    type: boolean
                created_at:
                    type: string
                    format: date-time
            required:
                - tweet_id
                - user_info
                - likes_count
                - retweets_count
                - replies_count
                - is_quote
                - is_reply
                - is_pinned
                - has_liked
                - has_retweeted
                - created_at
        Media:
            type: object
            properties:
                type:
                    type: string
                url:
                    type: string
            required:
                - type
                - url
        UserInfoWithoutBio:
            type: object
            properties:
                user_id:
                    type: string
                user_name:
                    type: string
                profile_image_url:
                    type: string
                    format: uri
                is_private:
                    type: boolean
                is_admin:
                    type: boolean
                is_following:
                    type: boolean
                is_followed:
                    type: boolean
                is_pending:
                    type: boolean
            required:
                - user_id
                - user_name
                - profile_image_url
                - is_private
                - is_admin
                - is_following
                - is_followed
                - is_pending